# src/main/resources/static/api-docs/swagger.yaml
openapi: 3.0.3
info:
  title: "Todo Cine API"
  description: "API managed with Spring Boot, JWT security, and movie catalog with custom exception handling."
  version: "v6.3.0"

servers:
  - url: "http://localhost:8080/api"
    description: "Local Server"

security:
  - BearerAuth: []

paths:
  /usuarios:

    post:
      summary: "Insert a new user"
      tags: ["Users"]
      security:
        - BearerAuth: []
      requestBody:
        description: "New user data"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioReqDTO'
      responses:
        '201':
          description: "Created - User registered successfully."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /usuarios/{id}:
    parameters:
      - name: "id"
        description: "Unique user ID"
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: "Get user by ID"
      tags: ["Users"]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: "OK - User found."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioDTO'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'

    put:
      summary: "Update user"
      tags: ["Users"]
      security:
        - BearerAuth: []
      requestBody:
        description: "Updated user data"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioReqDTO'
      responses:
        '200':
          description: "OK - User updated."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /usuarios/{userId}/movies:
    get:
      summary: "Get user movies (Paginated)"
      tags: ["User Movies"]
      security:
        - BearerAuth: []
      parameters:
        - name: "userId"
          description: "User ID"
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: "vista"
          description: "Filter by watched movies (true/false)"
          in: query
          required: true
          schema:
            type: string
        - name: "votada"
          description: "Filter by voted movies"
          in: query
          required: true
          schema:
            type: string
        - name: "orderBy"
          description: "Sorting field (e.g., popularity)"
          in: query
          required: true
          schema:
            type: string
        - name: "page"
          description: "Requested page number"
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: "OK - Paginated list retrieved."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paginator'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'

  /usuarios/{userId}/movies/{movieId}:
    parameters:
      - name: "userId"
        description: "User ID"
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: "movieId"
        description: "Movie ID"
        in: path
        required: true
        schema:
          type: integer
          format: int64
    put:
      summary: "Update movie status for a user"
      tags: ["User Movies"]
      security:
        - BearerAuth: []
      requestBody:
        description: "User-Movie relationship data"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioMovieDTO'
      responses:
        '200':
          description: "OK - Movie status updated."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetailDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'

  /movies:
    get:
      summary: "List movies with filters (Paginated)"
      tags: ["General Movies"]
      security:
        - BearerAuth: []
      parameters:
        - name: "name"
          description: "Movie title"
          in: query
          required: true
          schema: { type: string }
        - name: "status"
          description: "Status (Released, Post-Production...)"
          in: query
          required: true
          schema: { type: string }
        - name: "region"
          description: "Region code (ES, US...)"
          in: query
          required: true
          schema: { type: string }
        - name: "page"
          description: "Page number"
          in: query
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: "Filtered movie list."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paginator'
        '400':
          $ref: '#/components/responses/BadRequest'
        '502':
          $ref: '#/components/responses/BadGateway'

  /movies/{id}:
    get:
      summary: "Get full movie detail by ID"
      tags: ["General Movies"]
      security:
        - BearerAuth: []
      parameters:
        - name: "id"
          description: "Movie ID"
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: "Movie detail."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetailDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'

  /ganadores:
    get:
      summary: "Get winners by award and year"
      tags: ["Winners"]
      security:
        - BearerAuth: []
      parameters:
        - name: "premioId"
          description: "Award ID"
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: "anyo"
          description: "Award year"
          in: query
          required: true
          schema: { type: integer }
        - name: "pagina"
          description: "Results page"
          in: query
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: "List of winners retrieved."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paginator'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      summary: "Insert a new winner"
      description: "Requires ADMIN role."
      tags: ["Winners"]
      security:
        - BearerAuth: []
      requestBody:
        description: "Winner data"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GanadorReqDTO'
      responses:
        '201':
          description: "Winner registered."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GanadorDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
  /premios:
    get:
      summary: "Get awards"
      description: "Get awards"
      tags: [ "Awards" ]
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: "List of awards."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PremioDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'

  /premios/{id}:
    get:
      summary: "Get specific award"
      description: "Get award by unique identifier"
      tags: [ "Award" ]
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: "OK-Award."
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/PremioDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /premios/{id}/categorias:
    get:
      summary: "Get categories for a specific award"
      description: "Requires ADMIN role."
      tags: ["Awards & Categories"]
      security:
        - BearerAuth: []
      parameters:
        - name: "id"
          description: "Award ID (e.g., Oscars)"
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: "List of categories."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoriaDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'

  /usuarios/{userId}/listas:
      parameters:
        - name: "userId"
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: "Unique identifier of the user who owns the lists"

      get:
        summary: "Get user's lists of movies"
        description: "Returns a paginated list of all movie lists created by a specific user."
        tags: ["User's Lists"]
        security:
          - BearerAuth: [ ]
        parameters:
          - name: "page"
            in: query
            required: true
            schema:
              type: integer
            description: "Page index (1..N)"
        responses:
          '200':
            description: "Paginated lists retrieved successfully"
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Paginator'
          '400':
            $ref: '#/components/responses/BadRequest'
          '403':
            $ref: '#/components/responses/Forbidden'

      post:
        summary: "Create a new list of movies"
        description: "Registers a new list of movies associated with the specified user."
        tags: ["User's Lists"]
        security:
          - BearerAuth: [ ]
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListaReqDTO'
        responses:
          '201':
            description: "List created successfully"
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ListaDTO'
          '400':
            $ref: '#/components/responses/BadRequest'
          '403':
            $ref: '#/components/responses/Forbidden'

  /usuarios/{userId}/listas/{id}:
      parameters:
        - name: "userId"
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: "id"
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: "Unique identifier of the specific list of movies"

      get:
        summary: "Get details of the movie list"
        description: "Returns the full details of a specific list, including its metadata and movie collection."
        tags: ["User's Lists"]
        security:
          - BearerAuth: [ ]
        responses:
          '200':
            description: "List details retrieved successfully"
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ListaDTO'
          '400':
            $ref: '#/components/responses/BadRequest'
          '403':
            $ref: '#/components/responses/Forbidden'
          '404':
            $ref: '#/components/responses/NotFound'

      put:
        summary: "Update an existing list"
        description: "Updates the name, description, or the entire movie collection of a specific list."
        tags: ["User's Lists"]
        security:
          - BearerAuth: [ ]
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListaReqDTO'
        responses:
          '200':
            description: "List updated successfully"
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ListaDTO'
          '400':
            $ref: '#/components/responses/BadRequest'
          '403':
            $ref: '#/components/responses/Forbidden'
          '404':
            $ref: '#/components/responses/NotFound'

      delete:
        summary: "Delete a list of movies"
        description: "Permanently removes the list from the user's profile."
        tags: ["User's Lists"]
        security:
          - BearerAuth: [ ]
        responses:
          '204':
            description: "List deleted successfully (No Content)"
          '400':
            $ref: '#/components/responses/BadRequest'
          '403':
            $ref: '#/components/responses/Forbidden'
          '404':
            $ref: '#/components/responses/NotFound'

  /usuarios/{userId}/listas/{listaId}/movies/{movieId}:
      parameters:
        - name: "userId"
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: "listaId"
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: "movieId"
          in: path
          required: true
          schema:
            type: integer
            format: int64

      post:
        summary: "Add movie to list"
        description: "Adds a specific movie to the list. If the movie does not exist in the local database, it will be persisted."
        tags: ["User lists"]
        security:
          - BearerAuth: [ ]
        responses:
          '200':
            description: "Movie added successfully. Returns the updated list."
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ListaDTO'
          '400':
            $ref: '#/components/responses/BadRequest'
          '403':
            $ref: '#/components/responses/Forbidden'
          '404':
            $ref: '#/components/responses/NotFound'
          '502':
            $ref: '#/components/responses/BadGateway'

      delete:
        summary: "Remove movie from list"
        description: "Removes the relationship between the movie and the list without deleting the movie from the global catalog."
        tags: ["User's Lists"]
        security:
          - BearerAuth: [ ]
        responses:
          '204':
            description: "Movie removed from list successfully"
          '400':
            $ref: '#/components/responses/BadRequest'
          '403':
            $ref: '#/components/responses/Forbidden'
          '404':
            $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: "Invalid data."
    Conflict:
      description: "Resource already exists."
    Forbidden:
      description: "Access denied."
    NotFound:
      description: "Not found."
    BadGateway:
      description: "External server error."

  schemas:
    UsuarioReqDTO:
      type: object
      required: [username, password]
      properties:
        id: { type: integer, format: int64, description: "Unique identifier" }
        username: { type: string, description: "Access username" }
        password: { type: string, format: string, description: "Access password" }
        rol: { type: string, description: "Assigned role (USER/ADMIN)" }

    UsuarioDTO:
      type: object
      required: [username, rol]
      properties:
        id: { type: integer, format: int64, description: "Unique identifier" }
        username: { type: string, description: "Access username" }
        rol: { type: string, description: "Assigned role (USER/ADMIN)" }

    Paginator:
      type: object
      required: [page, total_pages, total_results]
      properties:
        page: { type: integer, description: "Current page" }
        results:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/MovieDTO'
              - $ref: '#/components/schemas/MovieDetailDTO'
              - $ref: '#/components/schemas/GanadorDTO'
          description: "List of results"
        total_pages: { type: integer, description: "Total pages" }
        total_results: { type: integer, description: "Total records" }

    MovieDetailDTO:
      type: object
      required: [id, title, original_title, poster_path, release_date, popularity, vote_count, vote_average, genres, original_language, total_votos_TC, votos_media_TC]
      properties:
        id: { type: integer, format: int64, description: "Movie ID" }
        original_title: { type: string, description: "Original title" }
        title: { type: string, description: "Translated title" }
        poster_path: { type: string, description: "Poster image path" }
        overview: { type: string, description: "Summary or synopsis" }
        release_date: { type: string, description: "Release date" }
        popularity: { type: number, format: double, description: "Popularity index" }
        vote_count: { type: integer, description: "Total external votes" }
        vote_average: { type: number, format: double, description: "Average external score" }
        original_language: { type: string, description: "Original language" }
        favoritos: { type: boolean, description: "Is marked as favorite by user" }
        voto: { type: number, format: double, description: "User's personal vote" }
        vista: { type: boolean, description: "Already watched by user" }
        total_votos_TC: { type: integer, description: "Total votes in Todo Cine" }
        votos_media_TC: { type: number, format: double, description: "Average score in Todo Cine" }
        genres:
          type: array
          items: { $ref: '#/components/schemas/GenreDTO' }
          description: "Genre list"
        videos:
          type: array
          items: { $ref: '#/components/schemas/VideoDTO' }
          description: "Trailers or video list"

    GenreDTO:
      type: object
      required: [id]
      properties:
        id: { type: string, description: "Genre ID" }
        name: { type: string, description: "Genre name" }

    VideoDTO:
      type: object
      required: [id, name, key, site, type]
      properties:
        id: { type: string, description: "Video identifier" }
        name: { type: string, description: "Video name" }
        key: { type: string, description: "Platform key (e.g., YouTube ID)" }
        site: { type: string, description: "Hosting site" }
        type: { type: string, description: "Video type (Trailer, Teaser)" }

    UsuarioMovieDTO:
      type: object
      required: [usuarioId, movieId, favoritos, vista]
      properties:
        usuarioId: { type: integer, format: int64, description: "User ID" }
        movieId: { type: integer, format: int64, description: "Movie ID" }
        favoritos: { type: boolean, description: "Favorite status" }
        vista: { type: boolean, description: "Watch status" }
        voto: { type: number, format: double, nullable: true, description: "Personal score" }

    MovieDTO:
      type: object
      required: [id, title, original_title, poster_path, release_date]
      properties:
        id: { type: integer, format: int64, description: "Unique ID" }
        original_title: { type: string, description: "Original title" }
        title: { type: string, description: "Title" }
        poster_path: { type: string, description: "Image path" }
        overview: { type: string, description: "Synopsis" }
        release_date: { type: string, description: "Release date" }
        popularity: { type: number, format: double, description: "Popularity" }
        vote_count: { type: integer, description: "Total votes" }
        vote_average: { type: number, format: double, description: "Average score" }
        original_language: { type: string, description: "Language" }
        genres:
          type: array
          items: { $ref: '#/components/schemas/GenreDTO' }
          description: "Associated genres"
        videos:
          type: array
          items: { $ref: '#/components/schemas/VideoDTO' }
          description: "Associated videos"
        total_votos_TC: { type: integer, description: "Local votes" }
        votos_media_TC: { type: number, format: double, description: "Local average" }

    GanadorReqDTO:
      type: object
      required: [premioId, categoriaId, anyo, movieId]
      properties:
        premioId: { type: integer, format: int64, description: "Award ID" }
        categoriaId: { type: integer, format: int64, description: "Category ID" }
        anyo: { type: integer, description: "Gala year" }
        movieId: { type: integer, format: int64, description: "Movie ID" }

    GanadorDTO:
      type: object
      required: [premioId, categoriaId, anyo, movieId]
      properties:
        premioId: { type: integer, format: int64, description: "Award ID" }
        premio: { type: string, description: "Award name" }
        categoriaId: { type: integer, format: int64, description: "Category ID" }
        categoria: { type: string, description: "Category name" }
        anyo: { type: integer, description: "Gala year" }
        movieId: { type: integer, format: int64, description: "Movie ID" }
        original_title: { type: string, description: "Original title" }
        title: { type: string, description: "Title" }
        poster_path: { type: string, description: "Poster image" }
        overview: { type: string, description: "Synopsis" }
        release_date: { type: string, description: "Release date" }

    CategoriaDTO:
      type: object
      required: [id, nombre]
      properties:
        id: { type: integer, format: int64, description: "Unique category ID" }
        nombre: { type: string, description: "Category description" }

    PremioDTO:
      type: object
      required: [ id, titulo ]
      properties:
        id: { type: integer, format: int64, description: "Unique award ID" }
        titulo: { type: string, description: "Award name" }
        anyos:
          type: array
          items: { type: integer}
          description: "Award years"

    ListaDTO:
      type: object
      required: [ id, nombre, descripcion, username ]
      properties:
        id: { type: integer, format: int64, description: "Unique list ID" }
        nombre: { type: string, description: "List name" }
        descripcion: { type: string, description: "List description" }
        username: { type: string, description: "List owner" }
        movies:
          type: array
          items: { $ref: '#/components/schemas/MovieDTO' }
          description: "Movies in the list"

    ListaReqDTO:
      type: object
      required: [ nombre, descripcion, username ]
      properties:
        id: { type: integer, format: int64, description: "Unique list ID" }
        nombre: { type: string, description: "List name" }
        descripcion: { type: string, description: "List description" }
        username: { type: string, description: "List owner" }
        movies:
          type: array
          items: { $ref: '#/components/schemas/MovieDTO' }
          description: "Movies in the list"